requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.2.0
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.1
distlib==0.3.8
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.3
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==23.5
django-health-check==3.18.1
django-import-export==3.3.8
django-ipware==6.0.4
django-jazzmin==2.6.0
django-log-viewer==1.1.7
django-redis==5.4.0
django-rest-swagger==2.2.0
django-structlog==7.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
drf-spectacular==0.27.1
et-xmlfile==1.1.0
ffmpeg==1.4
filelock==3.13.1
flake8==7.0.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupPy==1.14
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
psutil==5.9.8
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydub==0.25.1
pyflakes==3.2.0
PyJWT==1.7.1
pytest==8.0.1
pytest-cov==4.1.0
python-ipware==2.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
structlog==24.1.0
tablib==3.5.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
xlrd==2.0.1
xlwt==1.3.0